added Thin, Skip and WindowStat as impls of OnlineStat and refractored a bit
